Use a Result handling logic in application

+++ src/server-app/Application/Items/Commands/DeleteItem.cs
@@ -1,3 +1,4 @@
+using Application.Common.Core;
 using Application.Common.Interfaces;
 using MediatR;
 
@@ -5,12 +6,12 @@ namespace Application.Items.Commands;
 
 public class DeleteItem
 {
-    public class Command : IRequest<Unit>
+    public class Command : IRequest<Result<Unit>>
     {
         public int Id { get; set; }
     }
 
-    public class Handler : IRequestHandler<Command, Unit>
+    public class Handler : IRequestHandler<Command, Result<Unit>>
     {
         private readonly IApplicationDbContext _context;
 
@@ -19,15 +20,18 @@ public class DeleteItem
             _context = context;
         }
 
-        public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+        public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
         {
             var item = await _context.Items.FindAsync(request.Id);
+
+            if (item == null)
+                return null;
 
             _context.Items.Remove(item);
 
             await _context.SaveChangesAsync(cancellationToken);
 
-            return Unit.Value;
+            return Result<Unit>.Success(Unit.Value);
         }
     }
 }
